+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bugfix
 * fix flicker of dynamic prompt on enter [#757](https://github.com/jcubic/jquery.terminal/issues/757)
 * fix re-rendering of jQuery/DOM nodes [#759](https://github.com/jcubic/jquery.terminal/issues/759)
+* fix regression in less + images
 
 ## 2.32.1
 ### Bugfix